pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "meta-watcher"
version = "0.1.0"
description = "Cross-platform person-first desktop video overlay app powered by SAM 3.1"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
  { name = "OpenAI Codex" }
]
dependencies = [
  "numpy>=2.1",
  "Pillow>=11.0",
]

[project.optional-dependencies]
desktop = [
  "PySide6>=6.8",
  "opencv-python>=4.10",
  "PyYAML>=6.0",
]
mac = [
  "mlx-vlm>=0.4.3",
]
linux = [
  "torch>=2.7",
  "torchvision>=0.22",
  "transformers>=4.50",
  "accelerate>=1.5",
]
dev = [
  "pytest>=8.3",
]

[project.scripts]
meta-watcher = "meta_watcher.app:main"

[tool.setuptools.packages.find]
include = ["meta_watcher*"]
